+++ tests/eval_files/anyio_trio.py
@@ -1,3 +1,4 @@
+# type: ignore
 # ARG --enable-visitor-codes-regex=(TRIO220)
 # NOTRIO