+++ tests/eval_files/trio102_anyio.py
@@ -1,3 +1,4 @@
+# type: ignore
 import anyio
 from anyio import get_cancelled_exc_class