+++ tests/eval_files/trio104_anyio.py
@@ -1,3 +1,4 @@
+# type: ignore
 import anyio
 
 try: