+++ tests/eval_files/trio105_anyio.py
@@ -4,11 +4,11 @@ import anyio
 
 # nurseries don't exist in anyio, so don't error on it.
 async def foo():
-    nursery = anyio.open_nursery()
+    nursery = anyio.open_nursery()  # type: ignore
     await nursery.start()
     await nursery.start_foo()
 
     nursery.start()  # should not be triggered with anyio
-    None.start()
+    None.start()  # type: ignore
     nursery.start_soon()
     nursery.start_foo()