+++ tests/eval_files/trio110.py
@@ -1,3 +1,4 @@
+# type: ignore
 import trio
 import trio as noerror