+++ tests/eval_files/trio111.py
@@ -1,3 +1,4 @@
+# type: ignore
 from typing import Any
 
 import trio