+++ tests/eval_files/trio210.py
@@ -1,3 +1,4 @@
+# type: ignore
 import urllib
 
 import httpx