+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Future
 - Fix TRIO900 false alarm on nested functions
+- TRIO113 now also works on `anyio.TaskGroup`
 
 ## 23.2.3
 - Fix get_matching_call when passed a single string as base. Resolves possibly several false alarms, TRIO210 among them.